requirements.txt
django==4.2.3
djangorestframework==3.14.0
gunicorn==21.2.0
mysqlclient==1.4.6
python-dotenv==0.11.0